+++ cmem_plugin_reason/plugin_validate.py
@@ -182,9 +182,7 @@ class ValidatePlugin(WorkflowPlugin):
         for iri, filename in graphs.items():
             self.log.info(f"Fetching graph {iri}.")
             with (Path(self.temp) / filename).open("w", encoding="utf-8") as file:
-                if iri in missing and self.ignore_missing_imports:
-                    self.log.info(f"ignoring missing import {iri}.")
-                else:
+                if iri not in missing:
                     self.log.info(f"Fetching graph {iri}.")
                     setup_cmempy_user_access(self.context.user)
                     file.write(get(iri).text)
@@ -254,9 +252,7 @@ class ValidatePlugin(WorkflowPlugin):
         """Run the workflow operator."""
         setup_cmempy_user_access(self.context.user)
         graphs, missing = get_graphs_tree(
-            self,
-            graph_iris=(self.ontology_graph_iri, self.output_graph_iri),
-            ignore_missing=self.ignore_missing_imports,
+            self, graph_iris=(self.ontology_graph_iri,), ignore_missing=self.ignore_missing_imports
         )
         self.get_graphs(graphs, missing)
         create_xml_catalog_file(self.temp, graphs)